+++ tests/core/test_awgn_performance.py
@@ -1,11 +1,11 @@
 import unittest
 import numpy as np
-from src.comnumpy.core import Sequential
-from src.comnumpy.core.generators import SymbolGenerator
-from src.comnumpy.core.mappers import SymbolMapper, SymbolDemapper
-from src.comnumpy.core.utils import get_alphabet
-from src.comnumpy.core.channels import AWGN
-from src.comnumpy.core.metrics import compute_ser, compute_metric_awgn_theo
+from comnumpy.core import Sequential
+from comnumpy.core.generators import SymbolGenerator
+from comnumpy.core.mappers import SymbolMapper, SymbolDemapper
+from comnumpy.core.utils import get_alphabet
+from comnumpy.core.channels import AWGN
+from comnumpy.core.metrics import compute_ser, compute_metric_awgn_theo
 
 
 class TestAWGNChainPerformance(unittest.TestCase):